+++ src/module/vue/components/order-buttons.vue
@@ -1,0 +1,33 @@
+<template>
+  <div class="flexcol nogrow">
+    <i
+      class="clickable block fas fa-caret-up nogrow"
+      :class="{ disabled: i == 0 }"
+      @click="$emit('sortUp', i)"
+    ></i>
+    <i
+      class="clickable block fas fa-caret-down nogrow"
+      :class="{ disabled: i == length - 1 }"
+      @click="$emit('sortDown', i)"
+    ></i>
+  </div>
+</template>
+
+<style lang="less" scoped>
+div {
+  padding-right: 3px;
+
+  i {
+    padding: 2px;
+  }
+}
+</style>
+
+<script>
+export default {
+  props: {
+    i: Number,
+    length: Number,
+  },
+}
+</script>